pyproject.toml
["build-system"]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lucidscan"
version = "0.2.0"
description = "LucidScan - The trust layer for AI-assisted development"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "Voldeq GmbH", email = "toni.antunovic@voldeq.com" }
]
keywords = [
  "security",
  "scanner",
  "devsecops",
  "sast",
  "sca",
  "iac",
  "container",
  "vulnerability",
  "trivy",
  "semgrep",
  "checkov",
  "cli",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
  "Typing :: Typed",
]
dependencies = [
  "PyYAML>=6.0",
  "pathspec>=0.12.0",
  "langchain-core>=0.3",
  "langchain-openai>=0.2",
  "langchain-anthropic>=0.2",
  "langchain-ollama>=0.2",
  "questionary>=2.0",
  "Jinja2>=3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
lucidscan = "lucidscan.cli:main"

[project.entry-points."lucidscan.scanners"]
trivy = "lucidscan.plugins.scanners.trivy:TrivyScanner"
opengrep = "lucidscan.plugins.scanners.opengrep:OpenGrepScanner"
checkov = "lucidscan.plugins.scanners.checkov:CheckovScanner"

[project.entry-points."lucidscan.reporters"]
json = "lucidscan.plugins.reporters.json_reporter:JSONReporter"
table = "lucidscan.plugins.reporters.table_reporter:TableReporter"
summary = "lucidscan.plugins.reporters.summary_reporter:SummaryReporter"
sarif = "lucidscan.plugins.reporters.sarif_reporter:SARIFReporter"

[project.entry-points."lucidscan.enrichers"]
ai_explainer = "lucidscan.plugins.enrichers.ai_explainer:AIExplainerEnricher"

[project.entry-points."lucidscan.linters"]
ruff = "lucidscan.plugins.linters.ruff:RuffLinter"

# Default scanner plugin versions
# These are the source of truth for the build process
[tool.lucidscan.scanners]
trivy = "0.68.1"
opengrep = "1.12.1"
checkov = "3.2.495"
